# Conflicts:
#   Lab4/Lab4.X/nbproject/configurations.xml
#   Lab4/Lab4.X/nbproject/private/configurations.xml

+++ Lab4/Lab4.X/Part1_DCM2Euler.c
@@ -10,16 +10,16 @@
 #include <stdlib.h>
 #include <math.h>
 
-void DCM2Euler(float DCM[3][3], float Euler[3][1]) {
+void DCM2Euler(double DCM[3][3], double Euler[3][1]) {
     
     // solve for the yaw value in radians
-    float yaw_radians = atan2f(DCM[0][1], DCM[0][0]);
+    double yaw_radians = atan2f(DCM[0][1], DCM[0][0]);
     
     // solve for pitch in radians
-    float pitch_radians = -1*asinf(DCM[0][2]);
+    double pitch_radians = -1*asinf(DCM[0][2]);
     
     // solve for roll in radians
-    float roll_radians = atan2(DCM[1][2], DCM[2][2]);
+    double roll_radians = atan2(DCM[1][2], DCM[2][2]);
     
     // return the values in degrees
     Euler[0][0] = yaw_radians * 57.2958;
